Add a bunch of tests
Still don't have full coverage, but this change adds a lot of tests to
existing use cases. I had to refactor a lot to be able to write tests
properly so almost everything was touched.

Also added a default background to the main activity, so now it should
always show the app icon when launching the app